New stats system

+++ Assets/AOneButtonDefence/Scripts/Fighting/TargetManeqein.cs
@@ -6,7 +6,7 @@ public class TargetManeqein : MonoBehaviour, IDamagable
 
     public Transform GetTransform() => transform;
     
-    public void TakeDamage(IDamagable damagerTransform, int damage)
+    public void TakeDamage(IDamagable damagerTransform, float damage)
     {
         Debug.Log("Manequen attaked!");
     }
